Fix import/exports after converting package to ES module
